modify dir composition
